+++ package.json
@@ -17,8 +17,8 @@
     "node": ">=7.6"
   },
   "scripts": {
-    "test": "node node_modules/j1/run test",
-    "test_coverage": "istanbul cover node_modules/j1/run test"
+    "test": "node node_modules/j1/run test/*/index.js",
+    "test_coverage": "istanbul cover node_modules/j1/run test/*/index.js"
   },
   "dependencies": {
     "glob": "^7.1.1",
@@ -27,7 +27,6 @@
   },
   "devDependencies": {
     "istanbul": "^1.1.0-alpha.1",
-    "rollup": "^0.45.2",
-    "rollup-plugin-root-import": "^0.2.2"
+    "rollup": "^0.49.3"
   }
 }
