+++ test/.eslintrc.yml
@@ -1,4 +1,3 @@
 rules:
     no-console: off
-    global-require: off
     no-undef: off
