+++ test/001/src/deps1/a.js
@@ -1,0 +1,1 @@
+result.a = 'a';
