+++ test/001/src/deps1/b.js
@@ -1,0 +1,1 @@
+result.b = 'b';
