+++ test/001/src/deps2/c.js
@@ -1,0 +1,1 @@
+result.c = 'c';
