+++ test/001/src/deps2/d.js
@@ -1,0 +1,1 @@
+result.d = 'd';
