+++ test/001/src/index.js
@@ -1,0 +1,2 @@
+import './deps1/*.js';
+import '__dirname/deps2/*.js';
