--- test/projects/001/deps/002.js
@@ -1,1 +1,0 @@
-result.b = 3;
