--- test/projects/001/deps/003.js
@@ -1,1 +1,0 @@
-result.c = 4;
