--- test/projects/001/expected.json
@@ -1,5 +1,0 @@
-{
-  "a": 2,
-  "b": 3,
-  "c": 4
-}
