--- test/projects/001/index.js
@@ -1,1 +1,0 @@
-import './deps/*.js';
