--- test/projects/002/deps/001.js
@@ -1,1 +1,0 @@
-result.a = 2;
