--- test/projects/002/deps/002.js
@@ -1,1 +1,0 @@
-result.b = 3;
