--- test/projects/002/index.js
@@ -1,1 +1,0 @@
-import '__dirname/deps/*.js';
